// src/types/openapi/models/Customer.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Customer = {
    id?: number;
    firstName?: string | null;
    lastName?: string | null;
    address?: string | null;
    phone?: string | null;
    startDate?: string;
};