// src/types/openapi/models/SalesPerson.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SalesPerson = {
    id?: number;
    firstName: string;
    lastName: string;
    address: string;
    phone: string;
    startDate: string;
    terminationDate?: string | null;
    manager: string;
};